* libdb/db_lookup.c (make_content): Cast cont.dsize to size_t when comparing it with the return value of strlen(). Avoids warning when building with gdbm.

+++ libdb/db_lookup.c
@@ -273,7 +273,7 @@ datum make_content (struct mandata *in)
 		in->comp,
 		in->whatis);
 
-	assert (strlen (cont.dptr) + 1 == cont.dsize);
+	assert (strlen (cont.dptr) + 1 == (size_t) cont.dsize);
 #else /* !ANSI_SPRINTF */
 	sprintf (cont.dptr, "%s\t%s\t%s\t%ld\t%c\t%s\t%s\t%s\t%s",
 		 dash_if_unset (in->name),
